+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## evo-sdk@v0.1.12
+### What's changed
+#### evo-sdk
+* Add methods on DownloadedObject to download category data and attributes by @BenLewis-Seequent in https://github.com/SeequentEvo/evo-python-sdk/pull/124
+* Add MultiDownloadRequestHandler to test_tools, to be able to mock multiple downloads at the same time by @BenLewis-Seequent in https://github.com/SeequentEvo/evo-python-sdk/pull/124
+
+**Full changelog**: https://github.com/SeequentEvo/evo-python-sdk/compare/evo-sdk@v0.1.11...evo-sdk@v0.1.12
+
+## evo-objects@v0.3.2
+### What's changed
+#### evo-objects
+* Add methods on DownloadedObject to download category data and attributes by @BenLewis-Seequent in https://github.com/SeequentEvo/evo-python-sdk/pull/124
+
+**Full changelog**: https://github.com/SeequentEvo/evo-python-sdk/compare/evo-objects@v0.3.1...evo-objects@v0.3.2
+
+## evo-sdk-common@v0.5.8
+### What's changed
+#### evo-sdk-common
+* Add MultiDownloadRequestHandler to test_tools, to be able to mock multiple downloads at the same time by @BenLewis-Seequent in https://github.com/SeequentEvo/evo-python-sdk/pull/124
+
+**Full changelog**: https://github.com/SeequentEvo/evo-python-sdk/compare/evo-sdk-common@v0.5.7...evo-sdk-common@v0.5.8
+
 ## evo-sdk@v0.1.11
 ### What's changed
 #### evo-sdk